refactor(congressional-tech): finish youtube/congress package split
Complete the abandoned monorepo split so youtube-fetch/-analyze and
congress-fetch/-analyze work, without duplicating code.

- New shared package packages/congress_shared holds auth.py + globals.py
  (moved from apps/committee_youtube/src) plus the bundled committee CSV and
  congress metadata; globals.py resolves those relative to the package.
- Fix broken cross-package imports in youtube_api + congress_api:
  from apps.congress_youtube.{auth,globals} -> from congress_shared.{...}
  from packages.youtube_api.tables       -> from youtube_api.tables
- Declare the dependency graph with setuptools + path sources so an editable
  install resolves everything (congress_api had no build-system at all).
- Console scripts now live in the real packages (youtube_api, congress_api);
  installing the committee_youtube aggregator app puts all four on PATH.
- Fix two runtime bugs in youtube_api/tables.py: get_all_commitee_names now
  supports with_index=True, and get_committee_index passes csv_path through.
- Restore youtube-accounts.csv (from history) and congress_metadata.json
  (from apps/site) at the paths globals.py resolves to.
- Devcontainer installs the graph editable (python 3.12); update-youtube.yml
  paths corrected, PATH exported, schedule + pull_request re-enabled.
