+++ .devcontainer/devcontainer.json
@@ -9,7 +9,7 @@
       "nodeGypDependencies": true
     },
     "ghcr.io/devcontainers/features/python:1": {
-      "version": "3.11",
+      "version": "3.12",
       "installTools": true
     },
     "ghcr.io/devcontainers/features/git:1": {},
